% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{NxtIRFdata-package}
\alias{NxtIRFdata-package}
\alias{mock_genome()}
\alias{mock_gtf()}
\alias{example_bams()}
\alias{get_mappability_exclusion()}
\title{NxtIRFdata}
\description{
This package contains a workable example for the NxtIRF package.
It is based on an artificial chromosome Z,
as well as 6 example bam files based on samples from the Leucegene dataset
(GSE67039). The genes SRSF1, SRSF2, SRSF3, TRA2A, TRA2B, TP53 and NSUN5
sequences are used to construct chromosome Z, and the GTF denotes the
coordinates of these, based on Ensembl release 94. Bam files are constructed
based on the complete bam files of 6 samples from Leucegene. Bam files
are subset by regions of the 7 above genes, then the reads of these were
realigned to the mock reference using STAR.\cr\cr
Additionally, NxtIRFdata contains Mappability exclusion regions generated
using NxtIRF, suitable for use in generating referenced based on hg38,
hg19, mm10 and mm9 genomes.
}
\examples{
mock_genome() # returns the genome.fa file of the mock reference

mock_gtf() # returns the transcripts.gtf file of the mock reference

example_bams() # returns the locations of the 6 example bam files

get_mappability_exclusion("hg38") # Mappability exclusion BED for hg38
}
\keyword{package}
